[📝] Add docs for future external API
Add partial documentation for a planned API that is used via .json files and allows the addition of units, regions, and even custom actions. The API will be used by the vanilla game and will allow people wanting to make small changes to create 'mods' which let them do that.